Create a todo form component
`<TodoForm>` will hold an input field, `Add Todo` and
`Clear Completed` buttons.

The input field should take in user input, and allow a user to
press `Enter` or click on the `Submit Button` to add a todo to
your list.

Once a todo is submitted, the Todo List should re-render and show
the added todo.